requirements.txt
pandas
numpy
streamlit
openpyxl
seaborn
plotly-express